Rename internal plugins to remove “-config” sufix
