+++ src/plugins/webpack/base-plugin.js
