+++ src/plugins/webpack/json-plugin.js
