+++ src/plugins/webpack/pages-plugin.js
